Endor Han & Secret Objective
Triggers Custom Script when Han Strikes or when any copy of Secret
Objective is resolved in an Edge Stack.   Script prompts user to change
the objective and then offers a choice of all opponents objectives if
the user selects yes.  Programmed not to fire in Hardcore mode.

Additionally a Dev Script was added to add a card to your hand based on
GUID using the command call remoteCall(me,'addC',['<cardGUID>']). This
script was modified from the one db0 included in Netrunner.